+++ packages/headless/src/features/result-preview/result-preview-state.ts
@@ -1,0 +1,11 @@
+export interface ResultPreviewState {
+  uniqueId: string;
+  content: string;
+}
+
+export function getResultPreviewInitialState(): ResultPreviewState {
+  return {
+    uniqueId: '',
+    content: '',
+  };
+}